Forms: show countryname for flags
